Updates storybook
