+++ src/utils/utils.ts
@@ -44,3 +44,31 @@ export const getStorybookEnumValues = (enumObject, name) => {
   };
   return { options, getValue };
 };
+
+/**
+ * Given a pagination's pageCount, this will return (1) a page number,
+ * derived from the current URL, and (2) a function that, when passed to
+ * Pagination component, makes the URL change and refreshes the page.
+ * @NOTE this is only used for Storybook documentation.
+ */
+export const getStorybookHrefProps = (pageCount) => {
+  // This uses the `addon-queryparams` Storybook addon.
+  const urlParams = new URLSearchParams(document.location.search);
+  const pageParam = urlParams.get("page");
+
+  const getPageNumber = (page: number, pageCount: number) => {
+    return page > 0 && page <= pageCount ? page : 1;
+  };
+  const computedCurrentPage =
+    pageParam &&
+    Number(pageParam) &&
+    getPageNumber(Number(pageParam), pageCount);
+  const location = window.location;
+  // Passing this function into `Pagination` makes the URL to change
+  // and refreshes the page.
+  const getPageHref = (selectedPage) => {
+    return `${location.href}&page=${selectedPage}`;
+  };
+
+  return { computedCurrentPage, getPageHref };
+};
